Move ndcube reproject and rebin tests to own module.
